+++ src/main/java/io/reactivesocket/internal/PublisherUtils.java
@@ -27,6 +27,10 @@ import org.reactivestreams.Subscription;
 
 import io.reactivesocket.Frame;
 import io.reactivesocket.Payload;
+import io.reactivesocket.internal.rx.BackpressureHelper;
+import io.reactivesocket.internal.rx.BackpressureUtils;
+import io.reactivesocket.internal.rx.EmptySubscription;
+import io.reactivesocket.internal.rx.SubscriptionHelper;
 
 public class PublisherUtils {
