+++ src/main/java/io/reactivesocket/internal/Requester.java
@@ -38,6 +38,9 @@ import io.reactivesocket.Payload;
 import io.reactivesocket.exceptions.CancelException;
 import io.reactivesocket.exceptions.Exceptions;
 import io.reactivesocket.exceptions.Retryable;
+import io.reactivesocket.internal.rx.BackpressureUtils;
+import io.reactivesocket.internal.rx.EmptyDisposable;
+import io.reactivesocket.internal.rx.EmptySubscription;
 import io.reactivesocket.observable.Disposable;
 import io.reactivesocket.observable.Observer;
 import uk.co.real_logic.agrona.collections.Int2ObjectHashMap;
