+++ src/main/java/io/reactivesocket/internal/rx/BackpressureHelper.java
@@ -10,7 +10,7 @@
  * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See
  * the License for the specific language governing permissions and limitations under the License.
  */
-package io.reactivesocket.internal;
+package io.reactivesocket.internal.rx;
 
 import java.util.concurrent.atomic.*;
