+++ src/main/java/io/reactivesocket/internal/rx/BooleanDisposable.java
@@ -1,4 +1,4 @@
-package io.reactivesocket.internal;
+package io.reactivesocket.internal.rx;
 
 import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
