+++ src/main/java/io/reactivesocket/internal/rx/EmptyDisposable.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivesocket.internal;
+package io.reactivesocket.internal.rx;
 
 import io.reactivesocket.observable.Disposable;
