+++ src/main/java/io/reactivesocket/internal/rx/EmptySubscription.java
@@ -11,7 +11,7 @@
  * the License for the specific language governing permissions and limitations under the License.
  */
 
-package io.reactivesocket.internal;
+package io.reactivesocket.internal.rx;
 
 import org.reactivestreams.*;
