+++ src/main/java/io/reactivesocket/internal/rx/SubscriptionArbiter.java
@@ -30,9 +30,6 @@ import java.util.concurrent.atomic.*;
 
 import org.reactivestreams.Subscription;
 
-import io.reactivesocket.internal.BackpressureHelper;
-import io.reactivesocket.internal.SubscriptionHelper;
-
 /**
  * Arbitrates requests and cancellation between Subscriptions.
  */
